Python/src/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
beautifulsoup4
langchain
langchain-google-genai
requests